Refactor code to support the new ui
I refactored getHumanChoice() to initGame() which now accepts input from
the user via the added buttons. It consequently calls the callback
function playGame() which is responsible for playing each round by
calling playRound() as before. Now, playRound() no longer prints
messages to the console but directly to the screen now. The scores and
round is now directly reflected on the scoreboard.